chore: refactor store mixin
